Move message context header into same directory as other headers
